Initial commit
The git toolbox should satisfy all your git needs in terms of automations in
a transparent, non-intrusive way. In most situations, you should not even
notice that it exists while working on your project.

The git toolbox will be a pluggable system, a framework for interoperating
plugins.

The initial development will focus on showcasing its pluggability with the help
of a blogging engine as a static site generator, but as the API will stabilize,
it will be a good platform for managing and sharing your workflow in a more
generic way.

The blog engine generates blog articles from commit messages which are longer
than one line (i.e. which have a body).

This is the initial commit, and you are probably reading it as a "blog post".